Fix compilation errors: Counter.add() cast and Supplier constructor signature
- Cast int to long for Counter.add() calls (clientReadBytes, diskReadBytes, s3ReadBytes)
- Move StatsLogger initialization earlier in RemoteFileServer.start()
- Pass null for StatsLogger in RemoteRandomAccessReader.Supplier constructor calls

+++ herddb-remote-file-service/src/main/java/herddb/remote/RemoteRandomAccessReader.java
@@ -254,7 +254,7 @@ public class RemoteRandomAccessReader implements RandomAccessReader {
             clientReadLatency.registerSuccessfulEvent(elapsedNanos, TimeUnit.NANOSECONDS);
         }
         if (clientReadBytes != null) {
-            clientReadBytes.add(data.length);
+            clientReadBytes.add((long) data.length);
         }
         blockBuffer = data;
         bufferedBlockIndex = bufferIndex;
